Created my folders (added assets folder) and began website
Assests folder is essentially a file to which I uploaded the project I worked on last year in order to allow people to download the files directly. A suggestion I might have is to create one of these for everyone so they can do the same but if file upload / download speed is affected it might not be optimal.